+++ NEWS.md
@@ -1,3 +1,9 @@
+# jskm 0.5.5
+
+* Update: Add censoring marks to and `svyjskm`.
+
+* Update: Align color of censoring marks with color of lines in `jskm` and `svyjskm`.
+
 # jskm 0.5.4
 
 * Update: Add customization of line colors to `jskm` and `svyjskm`